move out of realize
